Remove driver.cc and mclass.cc
Those files don't do much in C++, so remove them for now.
