+++ core/module.cc
@@ -519,3 +519,30 @@ void dump_pcap_pkts(struct gate *gate, struct pkt_batch *batch) {
 }
 
 #endif
+
+size_t list_mclasses(const ModuleClass **p_arr, size_t arr_size,
+                     size_t offset) {
+  size_t ret = 0;
+  size_t iter_cnt = 0;
+
+  struct ns_iter iter;
+
+  ns_init_iterator(&iter, NS_TYPE_MCLASS);
+  while (1) {
+    ModuleClass *mc_obj = (ModuleClass *)ns_next(&iter);
+    if (!mc_obj) break;
+
+    if (iter_cnt++ < offset) continue;
+
+    if (ret >= arr_size) break;
+
+    p_arr[ret++] = mc_obj;
+  }
+  ns_release_iterator(&iter);
+
+  return ret;
+}
+
+const ModuleClass *find_mclass(const char *name) {
+  return (ModuleClass *)ns_lookup(NS_TYPE_MCLASS, name);
+}
