+++ core/port.cc
@@ -14,7 +14,7 @@
 #include "namespace.h"
 #include "port.h"
 
-Port::~Port() { ; }
+Port::~Port() {}
 
 size_t list_ports(const Port **p_arr, size_t arr_size, size_t offset) {
   size_t ret = 0;
@@ -337,4 +337,43 @@ void release_queues(Port *p, const struct module *m, packet_dir_t dir,
 
     if (p->users[dir][qid] == m) p->users[dir][qid] = NULL;
   }
+}
+
+size_t list_drivers(const Driver **p_arr, size_t arr_size, size_t offset) {
+  size_t ret = 0;
+  size_t iter_cnt = 0;
+
+  struct ns_iter iter;
+
+  ns_init_iterator(&iter, NS_TYPE_DRIVER);
+  while (1) {
+    Driver *driver = (Driver *)ns_next(&iter);
+    if (!driver) break;
+
+    if (iter_cnt++ < offset) continue;
+
+    if (ret >= arr_size) break;
+
+    p_arr[ret++] = driver;
+  }
+  ns_release_iterator(&iter);
+
+  return ret;
+}
+
+const Driver *find_driver(const char *name) {
+  return (Driver *)ns_lookup(NS_TYPE_DRIVER, name);
+}
+
+void init_drivers() {
+  struct ns_iter iter;
+
+  ns_init_iterator(&iter, NS_TYPE_DRIVER);
+  while (1) {
+    Driver *driver = (Driver *)ns_next(&iter);
+    if (!driver) break;
+
+    driver->Init();
+  }
+  ns_release_iterator(&iter);
 }
